+++ sc/social/like/tests/test_robot.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+from plone.testing import layered
+from sc.social.like.testing import ROBOT_TESTING
+
+import os
+import robotsuite
+import unittest
+
+
+dirname = os.path.dirname(__file__)
+files = os.listdir(dirname)
+tests = [f for f in files if f.startswith('test_') and f.endswith('.robot')]
+
+
+def test_suite():
+    suite = unittest.TestSuite()
+    suite.addTests([
+        layered(
+            robotsuite.RobotTestSuite(t, noncritical=['Expected Failure']),
+            layer=ROBOT_TESTING)
+        for t in tests
+    ])
+    return suite